code reactored for user service and decrypted the 2nd pubsub encrypted message to send to sendgrid and twilio
